+++ CHANGELOG.md
@@ -8,7 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Add `DataCube.aggregate_spatial()`
+
 ### Changed
+- Get/create default `RefreshTokenStore` lazily in `Connection`
 
 ### Removed